Moved the server to a different file
