+++ db.js
@@ -1,7 +1,7 @@
 const mongoose = require('mongoose')
 
 mongoose.connect(
-  'mongodb://127.0.0.1:27017/freelance',
+  proocess.env.MONGODB_URI,
     {
       useNewUrlParser: true,
       useUnifiedTopology: true,
